Replace Base.eltype with Quaternionic.basetype (#71)
This will cause breaking changes.  First, `Base.eltype` is no longer defined on any quaternionic type.  It was defined to return the `T` in an `AbstractQuaternion{T}`.  This was the source of some broadcasting issues, because broadcasting essentially recursively searches for the `eltype` of a vector, meaning that a vector of — say — `Quaternion{Float64}` would appear to broadcasting to have type `Float64`.  The new function `basetype` provides the same functionality.  This closes #15 and #70.

While fixing this, a floating-point change in the tests pointed out that `conj` and `-` acting on `Rotor`s would re-normalize, which is not really necessary and — in the case of `Float16` — can significantly change the components even when the original was already normalized.  These now skip the extra normalization step.

Changing that, in turn, uncovered a subtle bug in the method of `align` that was intended to align vectors of `Rotor`s, but did not actually specify the input types, allowing it to mistakenly apply to vectors that were intended to be `QuatVec`s, though one accidentally changed to `Quaternion`.  This change will cause a MethodError to raise when using unintended types.